Prepare campaign for C21_WMDE_Mobile_EN_Test_01
