+++ campaign_info.toml
@@ -49,24 +49,24 @@ orientation = [ "portrait", "landscape"]
 
 # Mobile English Campaign
 [mobile_english]
-campaign_tracking = "mob_en01-ba-201104"
-preview_link = "/wiki/Wikipedia:Hauptseite?devbanner={{banner}}&useskin=minerva&banner=B17WMDE_webpack_prototype"
-preview_url = 'https://de.m.wikipedia.org/wiki/Wikipedia:Hauptseite?banner={{PLACEHOLDER}}'
+campaign_tracking = "mob_en01-ba-211103"
+preview_link = "/wiki/Wikipedia:Main_Page?devbanner={{banner}}&useskin=minerva&banner=B17WMDE_webpack_prototype"
+preview_url = 'https://en.m.wikipedia.org/wiki/Wikipedia:Hauptseite?banner={{PLACEHOLDER}}'
 wrapper_template = "wikipedia_org"
 
 [mobile_english.banners.ctrl]
 filename = "./mobile_english/banner_ctrl.js"
-pagename = "B20WMDE_mob_en01_201104_ctrl"
-tracking = "org-mob_en01-201104-ctrl"
+pagename = "B21_WMDE_Mobile_EN_Test_01_ctrl"
+tracking = "org-mob_en01-211103-ctrl"
 
 [mobile_english.banners.var]
 filename = "./mobile_english/banner_var.js"
-pagename = "B20WMDE_mob_en01_201104_var"
-tracking = "org-mob_en01-201104-var"
+pagename = "B21_WMDE_Mobile_EN_Test_01_var"
+tracking = "org-mob_en01-211103-var"
 
 [mobile_english.test_matrix]
-device = [ 'samsung_s10', 'nexus_6', 'iphone_xs_max', 'iphone_5s', 'iphone_se', "iphone_8", "iphone_12_mini", "iphone_7_plus", "iphone_11_pro_max"]
-orientation = [ "portrait", "landscape"]
+device = [ 'samsung_s10', 'nexus_6', 'iphone_xs_max', 'iphone_5s', 'iphone_se', "iphone_8", "iphone_12_mini", "iphone_7_plus", "iphone_11_pro_max" ]
+orientation = [ "portrait", "landscape" ]
 
 
 # iPad campaign
