+++ mobile_english/banner_ctrl.js
@@ -1,5 +1,5 @@
 // eslint-disable-next-line no-unused-vars
-import style from './styles/styles_ctrl.pcss';
+import style from './styles/styles.pcss';
 
 import * as formatters from '../shared/number_formatter/en';
 import { createCampaignParameters } from '../shared/campaign_parameters';
@@ -9,14 +9,12 @@ import BannerPresenter from '../shared/banner_presenter';
 import Translations from '../shared/messages/en';
 import useOfFundsText from '../node_modules/fundraising-frontend-content/i18n/en_GB/data/use_of_funds_content.json';
 
-import Banner from './Banner';
-import Slides from './components/Slides';
-import BannerText from './components/BannerText';
+import Banner from './components/Banner';
+import DonationForm from './components/ui/form/DonationFormWithHeaders';
 
 import { createCampaignProjection } from '../shared/campaign_projection';
 import { createFormItems } from './form_items';
 import { LocalImpressionCount } from '../shared/local_impression_count';
-import DonationForm from './components/ui/form/DonationFormWithHeaders';
 
 const bannerContainer = document.getElementById( 'WMDE-Banner-Container' );
 const campaignParameters = createCampaignParameters();
@@ -25,7 +23,7 @@ const trackingIds = getTrackingIds( bannerContainer );
 const trackingData = createTrackingData( trackingIds.bannerName );
 const bannerPresenter = new BannerPresenter(
 	trackingData,
-	bannerContainer.dataset.delay || 7500,
+	bannerContainer.dataset.delay || 5000,
 	new LocalImpressionCount( trackingIds.bannerName )
 );
 
@@ -38,8 +36,6 @@ bannerPresenter.present(
 		campaignProjection,
 		formatters,
 		useOfFundsText,
-		bannerText: BannerText,
-		slides: Slides,
 		donationForm: DonationForm,
 		sliderAutoPlaySpeed: 5000,
 		translations: Translations,
