+++ mobile_english/components/MiniBanner.jsx
@@ -1,10 +1,22 @@
 import { h } from 'preact';
-import Infobox from '../../shared/components/ui/Infobox';
 import ProgressBar, { AmountToShowOnRight } from '../../shared/components/ui/ProgressBar';
 import * as PropTypes from 'prop-types';
+import Slides from '../Slides';
+import Slider from '../../shared/components/Slider';
 
 export default function MiniBanner( props ) {
 	const campaignProjection = props.campaignProjection;
+	const ProgressBarComponent = <ProgressBar
+		formatters={props.formatters}
+		daysLeft={campaignProjection.getRemainingDays()}
+		donationAmount={campaignProjection.getProjectedDonationSum()}
+		goalDonationSum={campaignProjection.goalDonationSum}
+		missingAmount={campaignProjection.getProjectedRemainingDonationSum()}
+		setStartAnimation={props.setStartAnimation}
+		animate={true}
+		amountToShowOnRight={AmountToShowOnRight.MISSING}
+	/>;
+
 	return <div className="mini-banner">
 		<div className="mini-banner__box">
 			<div className="mini-banner__content">
@@ -15,28 +27,15 @@ export default function MiniBanner( props ) {
 				</header>
 				<div className="close-button" onClick={props.onClose}/>
 
-				<Infobox
-					campaignParameters={props.campaignParameters}
-					campaignProjection={campaignProjection}
-					formatters={props.formatters}
-					bannerText={props.slides}
-					propsForText={{
-						formattedGoalDonationSumNumeric: props.formatters.millionFormatterNumeric( campaignProjection.goalDonationSum ),
-						progressBar: ( <ProgressBar
-							formatters={props.formatters}
-							daysLeft={campaignProjection.getRemainingDays()}
-							donationAmount={campaignProjection.getProjectedDonationSum()}
-							goalDonationSum={campaignProjection.goalDonationSum}
-							missingAmount={campaignProjection.getProjectedRemainingDonationSum()}
-							setStartAnimation={props.setStartAnimation}
-							animate={true}
-							amountToShowOnRight={AmountToShowOnRight.MISSING}
-						/> )
-					}}
+				<Slider
+					slides={ Slides( props.dynamicCampaignText, ProgressBarComponent ) }
+					onSlideChange={ props.onSlideChange }
+					registerAutoplay={ props.registerSliderAutoplayCallbacks }
+					interval={ props.sliderAutoPlaySpeed }
 				/>
 
 				<div className="mini-banner__tab">
-					<div className="mini-banner__tab-inner" onClick={props.onExpandFullpage}>
+					<div className="mini-banner__tab-inner" onClick={ props.onExpandFullpage }>
 						Donate now
 					</div>
 				</div>
