+++ mobile_english/styles/ProgressBar.pcss
@@ -1,8 +1,10 @@
 .wmde-banner {
 	.progress_bar {
-		/* remove this when campaign starts */
-		display: none;
-		/* display: flex; */
+		display: flex;
+		margin-top: 10px;
+	}
+
+	.fullpage-banner .progress_bar {
 		margin-right: 1.2em;
 		margin-left: 1.2em;
 	}
