--- mobile_english/styles/SelectGroup_var.pcss
@@ -1,241 +1,0 @@
-.wmde-banner {
-
-	.select-group {
-		display: flex;
-		flex-wrap: wrap;
-		justify-content: space-between;
-	}
-
-	.select-group__state {
-		display: block;
-		box-sizing: border-box;
-		font-family: inherit;
-		font-size: 24px;
-		font-weight: bold;
-
-		color: $color-black-light;
-		background-color: $color-gray-medium-bright;
-		border-radius: $form-element-border-radius;
-		padding: 18px 2px;
-		min-height: 54px;
-	}
-
-	.select-group__input:checked + .select-group__state {
-		background-color: $color-primary;
-		color: $color-white;
-		border-color: $color-gray-medium;
-	}
-
-	.select-group__input:disabled + .select-group__state {
-		color: $color-secondary-light;
-		border-color: $color-secondary-light;
-	}
-
-	.select-group__input {
-		position: absolute;
-		overflow: hidden;
-		height: 1px;
-		width: 1px;
-		clip: rect( 0 0 0 0 );
-		border: 0;
-		margin: -1px;
-		padding: 0;
-	}
-
-	.select-group__option > label {
-		box-sizing: border-box;
-		display: block;
-		margin: 0 0 10px;
-		outline: 0;
-		text-align: center;
-		line-height: 1;
-		cursor: pointer;
-
-		transition: background 100ms, color 100ms, border-color 100ms, box-shadow 100ms;
-
-		&:focus,
-		&:active {
-			.select-group__state {
-				background-color: $color-secondary-light;
-				color: $color-black;
-				border-color: $color-gray-medium;
-			}
-
-			&::placeholder {
-				color: $color-primary-lighter;
-			}
-		}
-	}
-
-	.select-group__custom-input {
-		width: 100%;
-		box-sizing: border-box;
-
-		border-radius: $form-element-border-radius;
-
-		min-height: 54px;
-
-		display: block;
-		background-color: $color-gray-medium-bright;
-		padding: 15px 2px;
-		height: 60px;
-		outline: 0;
-		cursor: pointer;
-		transition: background 100ms, color 100ms, border-color 100ms, box-shadow 100ms;
-
-		&--input {
-			width: 100%;
-			line-height: 1;
-			color: $color-black-light;
-			text-align: center;
-			font-weight: bold;
-			font-family: inherit;
-			font-size: 24px;
-
-			&:active,
-			&:focus,
-			&.selected-option {
-				color: $color-primary-lighter;
-				&::placeholder {
-					color: $color-primary-lighter;
-				}
-			}
-		}
-
-		&--euro-symbol {
-			display: block;
-			position: absolute;
-			width: 30px;
-			height: 30px;
-			top: 50%;
-			left: 0;
-			margin-top: -19px;
-			text-align: center;
-			line-height: 30px;
-			font-size: 24px;
-			color: $color-white;
-		}
-
-		&--value-entered,
-		&--focused {
-			background-color: $color-primary;
-			border-color: $color-gray-medium;
-			.select-group__custom-input--input,
-			.select-group__custom-input--euro-symbol {
-				color: $color-white;
-			}
-		}
-	}
-
-	.select-group__notice {
-		display: block;
-		font-size: 9pt;
-		line-height: 12pt;
-		margin-right: 0;
-		width: 100%;
-		padding: 0 0 15px 0;
-		color: $color-secondary-dark;
-	}
-
-	.select-group__errormessage {
-		display: none;
-		color: $color-black;
-		font-weight: bold;
-		text-align: left;
-		width: 100%;
-		margin-bottom: 0.2em;
-		font-size: 0.8em;
-	}
-
-	.select-group-container--with-error {
-		.select-group__erroricon {
-			position: relative;
-			padding-left: 26px;
-			display: block;
-
-			&:before {
-				content: $errorIcon;
-				width: 16px;
-				height: 16px;
-				position: absolute;
-				top: 0;
-				left: 6px;
-			}
-		}
-
-		.select-group__errormessage {
-			display: block;
-			color: $color-error;
-		}
-	}
-
-	/* Individual button sizes for each section */
-
-	.select-group-container--select-interval {
-		.select-group__option {
-			width: 100%;
-			flex: 1 1 auto;
-			@media ( min-width: $breakpoint_phone_l ) {
-				width: 31%;
-				flex: 0 1 31%;
-			}
-		}
-	}
-
-	.select-group-container--select-amount {
-		.select-group__option {
-			width: 49%;
-			flex: 0 1 49%;
-			@media ( min-width: $breakpoint_phone_s ) {
-				width: 31%;
-				flex: 0 1 31%;
-			}
-		}
-
-		.select-group__option--amount-other-input {
-			position: relative;
-		}
-	}
-
-	.select-group-container--select-payment-method {
-		.select-group__option {
-			width: 100%;
-			flex: 1 1 auto;
-			@media ( min-width: $breakpoint_phone_l ) {
-				width: 49%;
-				flex: 0 1 49%;
-			}
-		}
-		.select-group__state {
-			min-height: 70px;
-			padding-top: 25px;
-			padding-bottom: 25px;
-		}
-
-		.select-group__notice {
-			display: none;
-			&.select-group__notice--selected {
-				display: block;
-			}
-		}
-
-		/* Sofortüberweisung button */
-		.select-group__option:nth-child( 5 ) {
-			width: 100%;
-			flex: 1 1 auto;
-		}
-
-		.sms-box .select-group__option {
-			width: 100%;
-			flex: 1 1 auto;
-		}
-	}
-
-	.select-group-container--address-option {
-		margin-bottom: 30px;
-		.select-group__option {
-			width: 100%;
-			flex: 1 1 auto;
-		}
-	}
-}
