--- mobile_english/styles/Slides.pcss
@@ -1,52 +1,0 @@
-.wmde-banner {
-	.carousel-cell {
-		display: flex;
-		align-items: center;
-		justify-content: center;
-		padding: 0 17px 0 17px;
-		cursor: grab;
-		width: 100%;
-		height: auto;
-		box-sizing: border-box;
-		p {
-			font-weight: bold;
-			font-family: $banner-main-font-family;
-		}
-		.goal-headline {
-			margin: 0 0 5px 0;
-		}
-	}
-
-	.mini-banner-carousel {
-		display: flex;
-	}
-
-	.navigation-wrapper {
-		overflow: hidden;
-		#dots-navigation {
-			display: flex;
-			padding: 10px 0;
-			justify-content: center;
-		}
-
-		.dot {
-			border: 0;
-			width: 10px;
-			height: 10px;
-			background: $color-primary;
-			opacity: 0.25;
-			border-radius: 50%;
-			margin: 0 5px;
-			padding: 5px;
-			cursor: pointer;
-		}
-
-		.dot:focus {
-			outline: 0;
-		}
-
-		.dot--active {
-			opacity: 1;
-		}
-	}
-}
