+++ mobile_english/styles/styles.pcss
@@ -11,9 +11,9 @@
 @import './Banner.pcss';
 @import './MiniBanner.pcss';
 @import './FullpageBanner.pcss';
-@import './Slides.pcss';
+@import 'Slider.pcss';
 @import './DonationForm.pcss';
-@import './SelectGroup_var.pcss';
+@import './SelectGroup.pcss';
 @import './SmsBox.pcss';
 @import './BannerText.pcss';
 @import './TextHighlight.pcss';
