--- mobile_english/styles/styles_ctrl.pcss
@@ -1,21 +1,0 @@
-@import '../../shared/components/ui/use_of_funds/variables.pcss';
-@import 'variables.pcss';
-
-@import '../../shared/components/BannerTransition.pcss';
-@import '../../shared/components/FollowupTransition.pcss';
-@import '../../shared/components/ui/use_of_funds/CompanyBudgets.pcss';
-@import '../../shared/components/ui/use_of_funds/FundsModal.pcss';
-@import '../../shared/components/ui/use_of_funds/FundsContent.pcss';
-@import '../../shared/components/ui/use_of_funds/FundsDistributionAccordion.pcss';
-
-@import './Banner.pcss';
-@import './MiniBanner.pcss';
-@import './FullpageBanner.pcss';
-@import './Slides.pcss';
-@import './DonationForm.pcss';
-@import './SelectGroup.pcss';
-@import './SmsBox.pcss';
-@import './BannerText.pcss';
-@import './TextHighlight.pcss';
-@import './ProgressBar.pcss';
-@import './Footer.pcss';
